requirements.txt
psycopg2-binary==2.8.3

django==2.1.7
django-common-helpers==0.9.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.1.1
factory_boy==2.12.0